requirements_testing.txt
# requirements for running tests (on top of environment.yml/requirements.txt)
pytest!=4.6.0
pytest-cov
pytest-timeout
pytest-harvest
flake8
flake8-array-spacing
numpydoc
codespell
pydocstyle
check-manifest
twine
wheel

# mne-nirs specific requirements
# redundancy with docs as all doc notebooks are tested
mne-bids
seaborn
fooof
pysnirf2>=0.3.1
